Fix errors and add test.

+++ src/main/kotlin/org/it/zoo/junit5/coroutines/Timeout.kt
@@ -3,6 +3,6 @@ package org.it.zoo.junit5.coroutines
 import java.util.concurrent.TimeUnit
 
 annotation class Timeout(
-    val value: Int,
+    val value: Long,
     val timeUnit: TimeUnit = TimeUnit.MILLISECONDS
 )
